refactor(runtime): interpreters share their machinery instead of copying it
runtime/interpreting.rb holds what CommandInterpreter and
EntityInterpreter each carried a copy of — the traced step and the
declared-argument coercion walk — so the two can no longer drift. The
command interpreter's other two jobs move beside it as named modules:
command_interpreter/argument_gate.rb (unknown/absent refusals, the
correlation-key allowance) and command_interpreter/mutation_applier.rb
(set/append/arithmetic, entity elements, creation attributes). The
pipeline file now reads as the order of the steps and how a record is
addressed, nothing else.
